move token type into lexer
